Update APT-Caching Server Using Apt-Cacher NG
The primary objective of this configuration is to establish an APT (Advanced Package Tool) caching server on the "sp-temp-dbrestore" server. This server, acting as a caching proxy, will efficiently manage and distribute software packages, significantly improving package management on connected servers. A key aspect of this setup is to restrict internet access for clients to protect them from the internet while allowing limited access only to the caching server.